requirements.txt
slack-sdk
python-dotenv
slack-bolt
openai
google-api-python-client
google-auth
requests